Completed single thread Hello World!
This program will first obtain the first activate namespace of each NVME
controller and arrange these namespaces in a singly linked list. The
program will iterate through each of the namespace, print basic
information about the namespace to stdout. Then it will use the SPDK api
to send a write command to write "Hello World!" string to the namespace
and then send a read command from the LBA location in the namespace to
host memory. Then the program will print the content of the read buffer
to stdout.